test(#664): relocate config-schema-drift + 4 spawn command suites to integration (C6.1-spawn)
The last grandfathered real-spawn unit files cross a genuine process boundary:
- config-schema-drift spawns scripts/gen-config-schema.ts --check (drift gate)
- commands/events spawns the CLI for the cross-process events cursor
- commands/improve-cli-flags + improve-result-to-file spawn a real CLI for an
  uncontended improve DB
- commands/distill/distill-cli-flag spawns the CLI to assert the removed verb
  is rejected

Relocate all five into tests/integration/ (preserving the commands/ subpath,
fixing import depth + __dirname/import.meta.dir CLI-path resolution) and drop
them from ALLOWED_SPAWN, lowering UNIT_PURITY_BASELINE 19 -> 14. Repoint the
three isolation-allowlist entries (events, improve-result-to-file,
distill-cli-flag) to their new integration paths (isolation baseline unchanged).

This empties the real-spawn lane of the unit-purity ratchet (ALLOWED_SPAWN now
holds only completions.test.ts, addressed separately).
