+++ tests/integration/config-schema-drift.test.ts
